requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
av==16.0.1
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.2
distro==1.9.0
Django==6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
faster-whisper==1.2.1
filelock==3.20.1
flatbuffers==25.12.19
fsspec==2025.12.0
google-auth==2.45.0
google-genai==1.56.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
onnxruntime==1.23.2
packaging==25.0
pillow==12.0.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyreadline3==3.5.4
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
torchaudio==2.9.1+cpu
torchvision==0.24.1+cpu
tqdm==4.67.1
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
websockets==15.0.1
yt-dlp==2025.12.8
